Basic refactoring, 1st step towards being reusable
