+++ app/javascript/react/components/MetadataEntry/RorAutocomplete.js
@@ -1,90 +1,33 @@
-import React, {useCallback, useState} from 'react';
-import {useCombobox} from 'downshift';
-import {comboboxStyles, menuStyles} from './shared';
-import axios from "axios";
-import _debounce from 'lodash/debounce';
-import stringSimilarity from 'string-similarity';
+import React, {useEffect, useState} from 'react';
+import GenericAutocomplete from "./GenericAutocomplete";
 
+export default function RorAutocomplete() {
+  // in order to use this component, we need to track the state of the autocomplete text and the autocomplete id
+  // https://www.freecodecamp.org/news/what-is-lifting-state-up-in-react/ is a better functional example than the react docs
+  // which show lifting state in class components.  It's also simpler and clearer.
 
-export default function RorAutocomplete({acText, setAcText, setAcID, setAutoBlurred}) {
-  const [inputItems, setInputItems] = useState([]);
+  const [acText, setAcText] = useState();
+  const [acID, setAcID] = useState();
+  const [autoBlurred, setAutoBlurred] = useState(false);
 
-  // see https://stackoverflow.com/questions/36294134/lodash-debounce-with-react-input
-  const debounceFN = useCallback(_debounce(supplyLookupList, 500), []);
-
-  function supplyLookupList(qt) {
-    axios.get('https://api.ror.org/organizations', { params: {query: qt},
-      headers: {'Content-Type': 'application/json; charset=utf-8', Accept: 'application/json'} } )
-        .then((data) => {
-          if (data.status !== 200) {
-            console.log('Response failure not a 200 response');
-          }else{
-            const myList = data.data.items;
-            setInputItems(sortSimilarity(myList, stringItem, qt), );
-          }
-        });
-  }
-
-  function stringItem(item){
-    return (item?.name || '');
-  }
-
-  function sortSimilarity(list, itemToStringFN, typedValue){
-    for (const item of list) {
-      item.similarity = stringSimilarity.compareTwoStrings(itemToStringFN(item), typedValue);
-    }
-
-    list.sort((x, y) => (x.similarity < y.similarity) ? 1 : -1 );
-
-    return list;
-  }
-
-  const {
-    isOpen,
-    getToggleButtonProps,
-    getLabelProps,
-    getMenuProps,
-    getInputProps,
-    getComboboxProps,
-    highlightedIndex,
-    getItemProps,
-  } = useCombobox({
-    items: inputItems,
-    onInputValueChange: ({inputValue}) => {
-      setAcText(inputValue);
-      // only autocomplete with 3 or more characters so as not to waste queries
-      if (!inputValue || inputValue.length < 4){
-        setInputItems([],);
-        return;
-      }
-      debounceFN(inputValue);
-    },
-    onSelectedItemChange: ({selectedItem}) => {
-      setAcID(selectedItem.id);
-    },
-    itemToString: (item) => stringItem(item),
-  });
+  // do something when blurring from the autocomplete, passed up here
+  useEffect(() => {
+        if(autoBlurred) {
+          console.log('blurred away from input!');
+        };
+        setAutoBlurred(false);
+      }, [autoBlurred]);
 
   return (
-      <div>
-        <label {...getLabelProps()}>Choose an element:</label>
-        <div style={comboboxStyles} {...getComboboxProps()}>
-          <input className='c-input__text' {...getInputProps()} value={acText} onBlur={ () => { setAutoBlurred(true) } } /> &#x2705; &#x2753;
-        </div>
-        <ul {...getMenuProps()} style={menuStyles}>
-          {isOpen &&
-          inputItems.map((item, index) => (
-              <li
-                  style={
-                    highlightedIndex === index ? {backgroundColor: '#bde4ff', marginBottom: '0.5em'} : { marginBottom: '0.5em' }
-                  }
-                  key={item.id}
-                  {...getItemProps({item, index})}
-              >
-                {item.name}
-              </li>
-          ))}
-        </ul>
-      </div>
-  );
-};
+      <>
+        <GenericAutocomplete acText={acText || ''} setAcText={setAcText} setAcID={setAcID} setAutoBlurred={setAutoBlurred} />
+        {acID
+            ? ''
+            : <span title="Institution not found. Select it from the auto-complete list if it's available.">&#x2753;</span>
+        }<br/>
+        <p>Typed value is: {acText}</p>
+        <p>Selected ID is: {acID}</p>
+        <p>Blurred: {'' + autoBlurred}</p>
+      </>
+  )
+}
