+++ stash/stash_engine/app/views/stash_engine/pages/test.html.erb
@@ -3,7 +3,7 @@
 
 <h1>test page</h1>
 
-<%= react_component('components/MetadataEntry/WrapAutocomplete') %>
+<%= react_component('components/MetadataEntry/RorAutocomplete') %>
 
 <p>
   Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ipsum faucibus vitae aliquet nec ullamcorper sit amet. Ac turpis egestas maecenas pharetra convallis. Leo vel orci porta non pulvinar neque laoreet suspendisse interdum. Elementum tempus egestas sed sed risus pretium quam. Hendrerit gravida rutrum quisque non tellus orci ac. Mattis molestie a iaculis at erat pellentesque. Eget duis at tellus at urna condimentum mattis pellentesque. Vehicula ipsum a arcu cursus. Sodales neque sodales ut etiam sit. Velit aliquet sagittis id consectetur purus ut. Amet risus nullam eget felis. At volutpat diam ut venenatis tellus in metus vulputate eu. Ipsum dolor sit amet consectetur adipiscing elit ut aliquam. Aliquet risus feugiat in ante metus dictum at. Vitae elementum curabitur vitae nunc sed velit dignissim. Quam pellentesque nec nam aliquam sem et tortor consequat id. Odio pellentesque diam volutpat commodo sed egestas egestas. Enim lobortis scelerisque fermentum dui faucibus in ornare quam.
